8 byte alignment of malloc for sgi and hppa too
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@6768 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/GC/stack.c
@@ -1408,7 +1408,7 @@ void PrintRef(ref(Object) ref)
   if (ref) {
     if (inBetaHeap(ref) && isObject(ref) ){
       fprintf(output, ", is object");
-      if (IsPrototypeOfProcess((ref)->Proto)){
+      if (IsPrototypeOfProcess((int)((ref)->Proto))){
 	fprintf(output, ", proto ok: 0x%x (", 
 		(int)ref->Proto);
 	DescribeObject(ref);
